Move MacAddress into ascii::setting::data_types
